Revert "animator: docs (#621)"
This reverts commit 69593ea9edab2fdb0c852443a9a4936ad17fd14c.
