+++ src/Animations/services/constants/constants.js
@@ -10,7 +10,7 @@ const small = 'small';
 const medium = 'medium';
 const large = 'large';
 
-const percentages = [0, 10, 20, 30, 40, 50, 60, 70, 80, 90, 100];
+const percentages = [10, 20, 30, 40, 50, 60, 70, 80, 90, 100];
 const timings = [micro, small, medium, large];
 const directions = [top, bottom, left, right];
 const sequences = ['default', 'flip', 'reverse', 'reverse-flip'];
