+++ src/Animations/services/helpers/child-helper.js
@@ -24,7 +24,6 @@ class ChildHelper {
     const {index, reverseIndex} = this;
     return new ClassBuilder(this.data)
       .withTranslateWrapper()
-      .withDebug()
       .withClassName(this.childProps.childClassName)
       .withSequence(index, reverseIndex)
       .build();
@@ -32,6 +31,7 @@ class ChildHelper {
 
   getLayer2() {
     return new ClassBuilder(this.data)
+      .withDebug()
       .withChild()
       .withOpacity()
       .withScale()
