+++ stories/Animations/Example.scss
@@ -1,37 +1,10 @@
 .basic-div {
   background: beige;
   border: 1px solid black;
-  font-size: 16px;
-  padding: 5px 20px;
+  font-size: 20px;
+  padding: 5px;
 }
 
 .basic-wrapper {
-  display: flex;
   height: 30px;
-}
-
-.flex-parent {
-  display: flex;
-  width: 100%;
-}
-
-.flex-grow-1 {
-  flex-grow: 1;
-}
-
-.flex-grow-4 {
-  color: purple;
-  flex-grow: 10;
-}
-
-.shuki-inner {
-  border: 1px solid black;
-}
-
-.shuki-child {
-  flex-grow:1;
-}
-
-.shuki-wrapper {
-  display: flex;
 }
