+++ stories/Animations/ShowHideControls.js
@@ -30,7 +30,7 @@ class ShowHideControls extends React.Component {
     }
     const intervalId = setInterval(() => {
       this.toggle();
-    }, 5000);
+    }, 2000);
 
     this.setState({intervalId});
   }
@@ -44,12 +44,10 @@ class ShowHideControls extends React.Component {
     return (
       <Container>
         <Row>
-          <Col span="2">
-            Show / Hide
-            <ToggleSwitch checked={this.state.show}
+          <Col span="1"><ToggleSwitch checked={this.state.show}
                                       onChange={() => this.onChange()}/>
           </Col>
-          <Col span="10">
+          <Col span="11">
             {React.createElement(children.type, {...children.props, show: this.state.show}, children)}
           </Col>
         </Row>
